Add guidelines for guidelines (#9)
I feel that it would be good to have some amount of standardization
around the documents that we add to this repo, at least when it comes to
guidelines documents. These kinds of documents are basically souped up
"do"s and "don't"s, and so it seems that we can assign a structure to
them.