Remove extraneous structure; add working TOC (#39)
This repository is fairly sparse at the moment, so while I understand
that we will eventually want organization and consistency, I have two
thoughts: 1) everyone may have a different idea on the kinds of
documents that we want to write; and 2) everyone may have a different
writing style. So I'm not sure that it makes sense to enforce any kind
of structure yet. I think right now we ought to focus on making it as
easy as possible for anyone (within the org, at least) to contribute.

To that end I've flattened the file structure so that all documents are
now in a `docs/` directory. I've also removed the guidelines for
guidelines, since I'm not sure that pattern will be used for all
documents (and even as I've been adding more, I've noticed I'm not
following my own rules, so I don't want to keep updating them until
they're more set in stone).

I've also updated the README to be a bit friendlier and fixed the table
of contents in the README, as they have been broken so far (now they
just list all of the files in `docs/`). Also, I don't know that we need
to necessarily say that this repository is in draft. It will always
be living; I've highlighted that in the README as well.