Update pull request merge guidelines
The current pull request merge guidelines are a bit misleading in how
the default Squash & Merge commit message is generated. It has been
updated to clarify that the default message is configurable in the
repository settings, so it can differ between repositories.

The advice has been scaled back a bit as well. Given the different
circumstances, I wasn't sure what exactly to advise so I have removed
the suggestion to edit the message for now. I'd like us to discuss
standardizing on one of the default commit message choices, then we can
revisit this guideline.

A recommendation has been added about leaving the default title
unaltered though. Altering it could break some of our scripts.