Remove guideline about preventing console logging
Logs are helpful, we shouldn't avoid them. Many of our projects have
alternative logging abstractions which are preferred over using
`console.log` directly, so the lint rule referenced here is very useful
to us. But these alternative abstractions are not shared across the
organization, and their use isn't really relevant to security, so this
recommendation to avoid `console.log` doesn't seem like a good fit for
this document.