Remove auditability section
The guideline about third party systems having audit logs has been
removed. It needs more detail about what needs to be logged in order
to be useful, and typically we don't have a whole lot of systems to
choose from anyway, so we might have little choice in the matter.