Update e2e-testing.md
Discussed the use of the term 'should' at the start of test names.  Using 'should' at the start of test names is a long standing and accepted standard.  We should be more aligned with the tools and frameworks guidelines and recommendations that we use.

Even repeated test names within a test file using 'should' at the start of test names should be considered valid.