chore: add comprehensive testing guidelines (#99)
## Description
This PR introduces a documentation outlining MetaMask's approach to
testing, including unit, integration, and end-to-end (e2e) tests. This
document provides aims to provide contributors with a clear
understanding of MetaMask's testing practices, encouraging comprehensive
test coverage and contributing to the robustness of the application.

## Changes
* Added `test_guidelines.md` to the docs folder in the contributor docs
GitHub repo.
* The document includes sections on integration tests, unit tests, and
e2e tests, each detailing the purpose, guidelines, and recommended
practices for these types of tests.