Add guidelines for maintaining changelogs
Changelogs seem straightforward enough to maintain, but recent updates
in the `core` repo indicate that there may be misconceptions. The
guidelines in this PR aim to clarify the purpose of a changelog and how
to maintain them so that they are valuable for consumers at all times.

This is a prequisite for updating the release process in the `core`
monorepo but should prove valuable in many other MetaMask projects as
well.